Don't overwrite other class attribute assignments.
When a snabbdom virtual node has both a class as a component of the `sel` property AND a property in an `attrs` object, the `attrs` object class overwrites the `sel` subcomponent class.

https://github.com/paldepind/snabbdom/blob/master/modules/attributes.js#L26
`elm.setAttribute(key, cur);`

Here's an example of `attrs` class overwriting `sel` subcomponent class:
http://esnextb.in/?gist=0512c49748550e57039bce437225081e

To prevent overwriting I suggest adding a check to determine if `key` equals "class", and if so use `classList.add` rather than `setAttribute`. 

To accommodate a list of classes in the `attrs` class property, `split` the value on spaces and loop over the resulting array adding each item.