chore(test): enable tests again for old google chrome
Chrome 49 was disabled in the browserstack tests because of failing
tests. Trying those tests now again, only a single CSS variable test is
failing. As chrome 49 was the first chrome version to support css
variables it is very likely that this is a bug in the browser itself,
especially as chrome 50 works as expected. Thus this commits enables the
browserstack tests for chrome 50

ISSUES CLOSED: #468