chore(test): remove snabbdom benchmarks
The benchmark code was not yet updated to version 3 of snabbdom and
thus cannot be run at the moment. Further, it depends on faker, which is
not available any more. So for now, the benchmark code and all its
dependencies have been removed. In the future we should again add
benchmarks and can reference this commit for some code.