feat: add comparator support to thunk and improve typing and encapsulation
Make it possible to compare arguments to thunks with a comparator
function for custom notions of equality

ISSUES CLOSED: #143
BREAKING CHANGE:
The signature of thunk is changed. The second argument is now an
options object that can contain a key and a comparator function. The
type of thunk is much stronger and stricter. The vnode properties used
by thunks are now private and not part of the public types. Changing the
function to thunk is no longer supported and does not cause the thunk to
be rebuild.